Add HybridCharacter and apparently some natural stair handling
HybridCharacter is a RigidCharacterBody3D and a CharacterBody3D combined.
It switches between the two depending on collision data. Accidentally came
across stair handling which is now present on both HybridCharacter and
RigidCharacterBody3D when keeping the acceleration at 15.0 instead of 3.0
while in the air causes the body to climb 0.25 m steps